% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvgHydraulics.R
\name{AvgHydraulics}
\alias{AvgHydraulics}
\title{A function to execute the Geomorphic Approach hydraulic simulation routine}
\usage{
AvgHydraulics(
  S,
  wb,
  db,
  db_max = NULL,
  b_value = NULL,
  max_Q = 1,
  D84,
  xs_output = TRUE
)
}
\arguments{
\item{S}{channel gradient (m/m)}

\item{wb}{reach averaged bankfull width (m).}

\item{db}{reach averaged bankfull depth (m).}

\item{db_max}{Defaults to NULL. reach averaged maximum bankfull depth (m).  Specifying db_max is preferred to calculate 'b'.}

\item{max_Q}{maximum discharge (m3/s) to simualted WUA for.  Defaults to 1 m3/s.}

\item{D84}{grain size (mm)}

\item{xs_output}{Defaults to TRUE. An expression specifying whether to produce a .csv and .jpg of the simulated channel cross section.}

\item{b}{User-specified b-value. Defaults to NULL and calculated within model unless specified.}
}
\value{
.csv and .jpeg of channel cross section if specified

data frame of reach-averaged hydraulics
AvgHydraulics()
}
\description{
This function executes the hydraulic geometry simulator to evaluate reach-averaged depths and velocities
generated at flows less than bankfull conditions.
For more information about this model see: McParland et al. (2016) and Gronsdahl et al. (XXXX)
}
