\name{GIFT_checklist_raw}
\alias{GIFT_checklist_raw}
\title{Raw checklist}
\description{
Returns a list of checklists.
}
\usage{
GIFT_checklist_raw()
}
\arguments{
}
\details{
}
\value{
A data frame with 12 columns, if namesmatched = FALSE.
}
\details{}
\section{Note}{}

\examples{
ex <- GIFT_checklist_raw(list_ID = c(1,5), api = api)
}
