% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_coverage.R
\name{GIFT_coverage}
\alias{GIFT_coverage}
\title{Taxonomic and trait coverage in GIFT}
\usage{
GIFT_coverage(
  what = "taxonomic_coverage",
  taxon_name = "Embryophyta",
  trait_ID = "1.1.1",
  GIFT_version = "latest",
  api = "https://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{what}{Taxonomic or trait coverage.}

\item{taxon_name}{Name of the taxonomic group you want to retrieve.}

\item{trait_ID}{Identification number of the trait you want to retrieve.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}
}
\value{
A data frame with either taxonomic or trait coverage per
GIFT polygon.
}
\description{
Retrieve either taxonomic or trait coverage (for a given trait) per polygon.
}
\details{
The output has 5 columns:

\emph{entity_ID} - Identification number of GIFT polygons\cr
\emph{total} - species richness or trait coverage\cr
\emph{native} - native species or trait coverage\cr
\emph{naturalized} - naturalized species or trait coverage\cr
\emph{endemic_min} - endemic species or trait coverage
}
\examples{
\dontrun{
ex <- GIFT_coverage(what = "taxonomic_coverage", taxon_name = "Angiospermae")
ex2 <- GIFT_coverage(what = "trait_coverage", taxon_name = "Angiospermae",
trait_ID = "1.2.1")

}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=GIFT_traits_meta]{GIFT_traits_meta()}}
}
