\name{GIFT_env_meta_misc}
\alias{GIFT_env_meta_misc}
\title{Miscellaneous information about environment}
\description{
Returns metadata about environment.
}
\usage{
GIFT_env_meta_misc()
}
\arguments{
}
\details{
}
\value{
A data frame with X columns.
}
\details{}
\section{Note}{}

\examples{
ex <- GIFT_env_meta_misc()
}
