% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_individual.R
\name{GIFT_individual}
\alias{GIFT_individual}
\title{The main function for conditional fine-mapping for in transcriptome-wide association studies with individual-level data}
\usage{
GIFT_individual(
  X,
  Y,
  Zx,
  Zy,
  gene,
  pindex,
  maxiter = 1000,
  tol = 1e-04,
  pleio = 0,
  ncores = 1
)
}
\arguments{
\item{X}{Complex gene expression matrix, each column is the standardized specific expression.}

\item{Y}{Standarized trait vector.}

\item{Zx}{Standardized cis-genotype matrix from eQTL data.}

\item{Zy}{Standardized cis-genotype matrix from GWAS data.}

\item{gene}{The gene names vector.}

\item{pindex}{A vector with each element represents the number of cis-SNPs for each gene.}

\item{maxiter}{The maximum iteration, which can be determined by users.}

\item{tol}{The convergence tolerance of the absolute value of the difference between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{pleio}{The option of controlling the pleiotropy, which can be determined by users. If pleio is set to 0, analysis will perform without controlling any SNP; If pleio is set to 1, analysis will perform  controlling the top SNP; If pleio is set to 2, analysis will perform controlling the top two SNPs.}

\item{ncores}{The number of cores used in analysis. If the number of cores is greater than 1, analysis will perform with fast parallel computing. The function mclapply() depends on another R package "parallel" in Linux.}
}
\value{
A data frame including the causal effect estimates and p values for the gene-based test.
}
\description{
GIFT_individual applies a likelihood-based approach, accounting for the correlated cis-SNPs and genes in a region
}
