\name{GIFT_references}
\alias{GIFT_references}
\title{References metadata}
\description{
Returns metadata about the references used to build GIFT database.
}
\usage{
GIFT_references()
}
\arguments{
\item{restricted}{Boolean indicating whether you want access to restricted data.}
}
\details{
}
\value{
A data frame with 14 columns.
}
\details{}
\section{Note}{}

\examples{
ex <- GIFT_references(restricted = FALSE)
}
