% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_richness.R
\name{GIFT_richness}
\alias{GIFT_richness}
\title{Species richness and trait coverage in GIFT}
\usage{
GIFT_richness(
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}

\item{taxon_ID}{Identification number of the taxonomic group you want to
retrieve.}

\item{trait_ID}{Identification number of the trait you want to retrieve.}

\item{what}{Species richness or trait coverage.}
}
\value{
data frame.
}
\description{
Retrieve environmental data associated to each GIFT checklists.
They can come as rasters or shapefiles (miscellaneous)
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_env()

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()]
}
