% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_shape.R
\name{GIFT_shape}
\alias{GIFT_shape}
\title{GIFT shape}
\usage{
GIFT_shape(
  entity_ID = NULL,
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{entity_ID}{A vector defining the IDs of the regions}
}
\value{
a spatial data.frame
}
\description{
Get shapefile of GIFT regions for selected entities
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_shape(entity_ID = c(677, 200))
plot(st_geometry(geodata), col=geodata$entity_ID)
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_plot()], [GIFT::GIFT_env()]
}
