% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_species_lookup.R
\name{GIFT_species_lookup}
\alias{GIFT_species_lookup}
\title{Species list in GIFT}
\usage{
GIFT_species_lookup(
  genus = "",
  epithet = "",
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{genus}{character string defining the genus name to be looked for.}

\item{epithet}{character string defining the specific epithet to be looked
for.}

\item{api}{character string defining from which API the data will be
retrieved.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data frame with 24 columns.
}
\description{
Retrieve all name matching information for one taxonomic name. All results 
are returned, where the name is either found in the unstandardized or 
taxonomically standardized names.
}
\details{
Here is what each column refers to:
'orig_ID' - Identification number of the species before taxonomic
harmonization
'orig_genus' - Identification number of the genus before taxonomic
harmonization
'name_ID' - Genus name before taxonomic harmonization
'cf_genus'- 
'genus'- Genus before taxonomic harmonization
'cf_species'- 
'aff_species'-
'species_epithet'- Epithet of the species before taxonomic harmonization
'subtaxon'- Subtaxon of the species before taxonomic harmonization
'author'- Author who described the species (before taxonomic harmonization)
'matched'- Is the species matched in the taxonomic backbone
'epithetscore'- Matching score for the epithet
'overallscore'- Overall matching score for the species
'resolved'- Is the species name resolved in the taxonomic backbone
'synonym'- Is the species name a synonym in the taxonomic backbone
'matched_subtaxon'- Is the subtaxon matched in the taxonomic backbone
'accepted'- Is the species name accepted in the taxonomic backbone
'service'- Service use for the taxonomic harmonization
'work_ID'- Identification number of the species after taxonomic
harmonization
'taxon_ID'- Identification number of the taxonomic group
'work_genus'- Identification number of the genus after taxonomic
harmonization
'work_species_epithet'- Identification number of the species epithet
after taxonomic harmonization
'work_species'- Species name (after taxonomic harmonization)
'work_author'-  Author who described the species (after taxonomic
harmonization)
}
\examples{
\dontrun{
ex <- GIFT_species_lookup(genus = "Fagus", epithet = "sylvatica")
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist()]
}
