% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_taxonomy.R
\name{GIFT_taxonomy}
\alias{GIFT_taxonomy}
\title{Taxonomy of GIFT}
\usage{
GIFT_taxonomy(
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}
}
\value{
A data frame with 6 columns.
}
\description{
Retrieves the taxonomy of GIFT.
}
\details{
Column \emph{taxon_ID} indicates the identification number of each
taxonomic entry. The names describing taxa are in the column
\emph{taxon_name}.The third column spells the author name for a given taxon.
The column \emph{taxon_lvl} splits every taxon in genus, family, order or
superior orders. Taxonomy is a linear sequence of left and right borders
for each taxon. This is nested, for example left and right borders of a
genus would fall between the left and right borders of the corresponding
family. Columns \emph{lft} and \emph{rgt} respectively refer to these left
and right borders.
}
\examples{
\dontrun{
ex <- GIFT_taxonomy()
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=GIFT_checklist]{GIFT_checklist()}}
}
