% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits.R
\name{GIFT_traits}
\alias{GIFT_traits}
\title{GIFT_traits}
\usage{
GIFT_traits(
  trait_IDs = "",
  agreement = 0.66,
  bias_ref = FALSE,
  bias_deriv = FALSE,
  api = "http://gift.uni-goettingen.de/api/extended/index.php"
)
}
\arguments{
\item{trait_IDs}{a character string indicating which trait you want to
retrieve. Traits must belong to the available list of traits.}

\item{agreement}{Traits desired.}

\item{bias_ref}{Traits desired.}

\item{bias_deriv}{Traits desired.}

\item{api}{Character string with the API.}
}
\value{
A long-format data frame with 6 columns: `trait_ID`, `work_ID`, `species`,
`trait_value`, `agreement` and `references`.
}
\description{
Retrieve specific trait values.
}
\details{
Blabla.
}
\examples{
\dontrun{
wood <- GIFT_traits(trait_IDs = c("1.1.1", "1.2.1"), agreement = 0.66,
bias_ref = FALSE, bias_deriv = FALSE)

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()]
}
