% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits_raw.R
\name{GIFT_traits_raw}
\alias{GIFT_traits_raw}
\title{GIFT_traits_raw}
\usage{
GIFT_traits_raw(
  trait_IDs = "",
  derived = TRUE,
  bias_ref = TRUE,
  bias_deriv = TRUE,
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{trait_IDs}{a character string indicating which traits you want to
retrieve. Traits must belong to the available list of traits. See 
GIFT_traits_meta()}

\item{derived}{include logically derived traits.}

\item{bias_ref}{When FALSE, exclude entries that are only based on a
resource that potentially introduces a bias (e.g. a resource only including
trees).}

\item{bias_deriv}{When FALSE, exclude entries that are only based on a
derivation that potentially introduces a bias (e.g. all phanerophytes being
woody but some life forms being ambiguous).}

\item{api}{Character string with the API.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data.frame with 26 columns.
}
\description{
Retrieve non-aggregated trait values at the level of the bibliographic 
references and un-standardized species names in GIFT.
}
\details{
Here is the detail of each column:
trait_derived_ID -
ref_ID -
orig_ID -
trait_ID -
trait_value -
derived -
bias_deriv -
name_ID -
cf_genus -
genus -
}
\examples{
\dontrun{
succulence_carnivory <- GIFT_traits_raw(trait_IDs = c("4.10.1", "4.16.1"))
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()] and [GIFT::GIFT_traits()]
}
