% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_two_stage_summ.R
\name{GIFT_two_stage_summ}
\alias{GIFT_two_stage_summ}
\title{The two-stage version of GIFT for conditional fine-mapping in transcriptome-wide association studies with existing gene expression prediction models}
\usage{
GIFT_two_stage_summ(betax, betay, se_betay, Sigma, n, gene)
}
\arguments{
\item{betax}{Weight matrix for the cis-SNP effect size.}

\item{betay}{Beta vector for the cis-SNP effect size from GWAS data.}

\item{se_betay}{Corresponding se vector for the cis-SNP effect size vector from GWAS data.}

\item{Sigma}{LD matrix from GWAS data.}

\item{n}{Sample size of GWAS data.}

\item{gene}{The gene names vector.}
}
\value{
A data frame including the z scores and p values for the gene-based test.
}
\description{
GIFT_two_stage_summ applies a two-stage approach, accounting for the correlated cis-SNPs and genes in a region.
}
