% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_version.R
\name{GIFT_version}
\alias{GIFT_version}
\title{Version of GIFT available}
\usage{
GIFT_version()
}
\value{
A data frame with 4 columns.
}
\description{
Returns a table with information on the different versions of the database
}
\details{
Here is what each column refers to:

\emph{ID} - Identification number of the version\cr
\emph{version} - Version number\cr
\emph{description} - What were the major updates about
}
\examples{
\dontrun{
ex <- GIFT_version()
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=GIFT_checklist]{GIFT_checklist()}}
}
