% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Habitat.R
\name{Habitat}
\alias{Habitat}
\title{A function to execute the ASHGS model}
\usage{
Habitat(
  hydraulics,
  max_Q = 10,
  d_curve,
  v_curve,
  s_curve = NULL,
  gsd = NULL,
  wua_output = TRUE
)
}
\arguments{
\item{hydraulics}{dataframe output generated by the AvgHydraulics function}

\item{max_Q}{maximum discharge (m3/s) to simualted WUA for.  Defaults to 1 m3/s.}

\item{d_curve}{dataframe of depth suitabilities with two columns: 'depth' (m) and 'suit' (dimensionless suitability from 0-1)}

\item{v_curve}{dataframe of velocity suitabilities with two columns: 'velocity' (m) and 'suit' (dimensionless suitability from 0-1)}

\item{s_curve}{dataframe of substrate suitabilities with three columns: 'lower' (mm) and 'upper' (mm) of grain size classes and 'suit' (dimensionless suitability from 0-1)}

\item{gsd}{vector of grain size distribution (mm)}

\item{wua_output}{defaults to TRUE. Expression to specify a .jpeg output of streamflow-WUA relationship}
}
\value{
data frame of streamflow, suitabilities, channel width, and WUA
Habitat()
}
\description{
This function executes the hydraulic geometry simulator to evaluate reach-averaged depths and velocities
generated at flows less than bankfull conditions.
For more information about this model see: McParland et al. (2016) and Gronsdahl et al. (XXXX)
}
