% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.R
\docType{data}
\name{med}
\alias{med}
\title{Shapefile of Mediterreanean region}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data("med")
}
\description{
Shapefile encompassing the Mediterreanean sea. CRS is WGS84.
}
\examples{
data(med)
}
\keyword{datasets}
