% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process_individual}
\alias{pre_process_individual}
\title{The pro-process function converts common genotype formats to match GIFT input using individual-level data.}
\usage{
pre_process_individual(filelocation, plinkexe = "plink")
}
\arguments{
\item{filelocation}{A file location only contains the genotype files to be processed.}

\item{plinkexe}{A path for the executable plink software.}
}
\value{
A list including the gene names, z-scores and cis-SNP numbers for each gene.
}
\description{
pre_process_individual is flexible to handle plink binary format (.bim/.fam./.bed), vcf, ped/map format, csv, and tsv file.
}
