% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process_summary}
\alias{pre_process_summary}
\title{The pro-process function converts common summary statistics and LD matrix format to match GIFT input using summary statistics.}
\usage{
pre_process_summary(
  eQTLfilelocation,
  eQTLLDfile,
  GWASfile,
  GWASLDfile,
  snplist,
  pindex
)
}
\arguments{
\item{eQTLfilelocation}{A file location only contains the summary statistics files from eQTL data.}

\item{eQTLLDfile}{A file path for the LD matrix from eQTL data.}

\item{GWASfile}{A file path for the summary statistics from GWAS data.}

\item{GWASLDfile}{A file path for the LD matrix from GWAS data.}

\item{snplist}{A vector represents the cis-SNP list for all genes in a region.}

\item{pindex}{A vector with each element represents the number of cis-SNPs for each gene.}
}
\value{
A list including the gene names, cis-SNP numbers for each gene, z-scores from eQTL data and GWAS data, LD matrix from eQTL data and GWAS data.
}
\description{
pre_process_summary flexible to handle association test output from plink (.qassoc), GEMMA (.assoc.txt) and SAIGE (.txt); and LD matrix either from matrix or a long format such as h5 format.
}
