% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process_twostage}
\alias{pre_process_twostage}
\title{The pro-process function converts common summary statistics and LD matrix format from GWAS data to match a required input of two-stage version of GIFT.}
\usage{
pre_process_twostage(GWASfile, GWASLDfile, snplist)
}
\arguments{
\item{GWASfile}{A file path for the summary statistics from GWAS data.}

\item{GWASLDfile}{A file path for the LD matrix from GWAS data.}

\item{snplist}{A vector represents the cis-SNP list for all genes in a region.}
}
\value{
A list including the beta vector, corresponding se vector and LD matrix from GWAS data.
}
\description{
pre_process_twostage flexible to handle association test output from plink (.qassoc), GEMMA (.assoc.txt) and SAIGE (.txt); and LD matrix either from matrix or a long format such as h5 format.
}
