% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{weightconvert}
\alias{weightconvert}
\title{The pro-process function converts the weights from all genes into a required input of two-stage version of GIFT.}
\usage{
weightconvert(weightlist)
}
\arguments{
\item{weightlist}{A list contains the weights for all genes in a region.}
}
\value{
A weight matrix to match the two-stage version of GIFT input.
}
\description{
weightconvert converts a list including the weights for multiple genes into a required block diagonal matrix.
}
