% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/western_mediterranean.R
\docType{data}
\name{western_mediterranean}
\alias{western_mediterranean}
\title{Shape file of the western Mediterranean basin}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data("western_mediterranean")
}
\description{
Shape file encompassing the western part of the Mediterranean basin.
CRS is WGS84.
}
\examples{
data(western_mediterranean)
}
\keyword{datasets}
