% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_checklist.R
\name{GIFT_checklist}
\alias{GIFT_checklist}
\title{GIFT checklists}
\usage{
GIFT_checklist(
  taxon_name = "Tracheophyta",
  complete_taxon = TRUE,
  floristic_group = c("all", "native", "endemic", "naturalized")[2],
  complete_floristic = TRUE,
  geo_type = c("Mainland", "Island"),
  suit_geo = FALSE,
  shp = NULL,
  coordinates = NULL,
  overlap = "centroid_inside",
  namesmatched = FALSE,
  GIFT_version = NULL,
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{taxon_name}{Character string corresponding to the taxonomic group
of interest.}

\item{complete_taxon}{Boolean, default TRUE.}

\item{floristic_group}{Character among the following options: 'all',
'native', 'endemic', 'naturalized'}

\item{complete_floristic}{Boolean, default TRUE.}

\item{geo_type}{Character, either 'Mainland' or 'Island'.}

\item{suit_geo}{Boolean.}

\item{shp}{Shapefile provided by the user.}

\item{coordinates}{Custom set of coordinates. The format is a two columns,
the first one being longitudes and the second being latitudes. If 4
coordinates are given, the function assumes that these are the four corners
of a bounding box.}

\item{overlap}{character vector or list defining the raster
data to retrieve..}

\item{namesmatched}{Logical}

\item{api}{character string defining from which API the data will be retrieved.}
}
\value{
List with two elements: the checklist with species and the list of ID.
}
\description{
Retrieve GIFT checklists that fulfill specific criteria.
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_checklist()
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist_raw()]
}
