% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_checklist_conditional.R
\name{GIFT_checklist_conditional}
\alias{GIFT_checklist_conditional}
\title{GIFT checklists}
\usage{
GIFT_checklist_conditional(
  taxon_name = "Tracheophyta",
  ref_included = c("all", "native", "native and naturalized",
    "native and historically introduced", "endangered", "endemic", "naturalized",
    "other subset")[1:4],
  type_ref = c("Account", "Catalogue", "Checklist", "Flora", "Herbarium collection",
    "Key", "Red list", "Report", "Species Database", "Survey"),
  entity_class = c("Island", "Island/Mainland", "Mainland", "Island Group",
    "Island Part"),
  native_indicated = FALSE,
  natural_indicated = FALSE,
  end_ref = FALSE,
  end_list = FALSE,
  suit_geo = TRUE,
  complete_taxon = TRUE
)
}
\arguments{
\item{taxon_name}{Character string corresponding to the taxonomic group
of interest.}

\item{ref_included}{Character, options are 'all', 'native',
'native and naturalized', 'native and historically introduced',
'endangered', 'endemic', 'naturalized', 'other subset'}

\item{type_ref}{Character, options are 'Account', 'Catalogue', 'Checklist',
'Flora', 'Herbarium collection', 'Key', 'Red list', 'Report',
'Species Database', 'Survey'}

\item{entity_class}{Character, options are 'Island', 'Island/Mainland',
'Mainland', 'Island Group', 'Island Part'}

\item{native_indicated}{Boolean.}

\item{natural_indicated}{Boolean.}

\item{end_ref}{Boolean.}

\item{end_list}{Boolean.}

\item{suit_geo}{Boolean.}

\item{complete_taxon}{Boolean, default TRUE.}
}
\value{
data frame with list of entity_ID and the different criteria.
}
\description{
Retrieve GIFT checklists that fulfill specific criteria.
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_checklist_conditional(taxon_name = "Embryophyta", 
ref_included = c("all", "native", "native and naturalized",
"native and historically introduced", "endangered",
"endemic", "naturalized", "other subset")[1:4],
type_ref = c("Account", "Catalogue", "Checklist","Flora",
"Herbarium collection", "Key", "Red list", "Report", "Species Database",
 "Survey"),
 entity_class = c("Island", "Island/Mainland", "Mainland", "Island Group",
 "Island Part"),
 native_indicated = FALSE, natural_indicated = FALSE, end_ref = FALSE,
 end_list = FALSE, suit_geo = TRUE, complete_taxon = TRUE)

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist_raw()]
}
