% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_checklist_conditional.R
\name{GIFT_checklist_conditional}
\alias{GIFT_checklist_conditional}
\title{GIFT checklists}
\usage{
GIFT_checklist_conditional(
  taxon_name = "Tracheophyta",
  ref_included = c("all", "native", "native and naturalized",
    "native and historically introduced", "endangered", "endemic", "naturalized",
    "other subset")[1:4],
  type_ref = c("Account", "Catalogue", "Checklist", "Flora", "Herbarium collection",
    "Key", "Red list", "Report", "Species Database", "Survey"),
  entity_class = c("Island", "Island/Mainland", "Mainland", "Island Group",
    "Island Part"),
  native_indicated = FALSE,
  natural_indicated = FALSE,
  end_ref = FALSE,
  end_list = FALSE,
  suit_geo = FALSE,
  complete_taxon = TRUE,
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/",
  list_set = NULL,
  taxonomy = NULL
)
}
\arguments{
\item{taxon_name}{Character string corresponding to the taxonomic group
of interest.}

\item{ref_included}{Character, options are 'all', 'native',
'native and naturalized', 'native and historically introduced',
'endangered', 'endemic', 'naturalized', 'other subset'}

\item{type_ref}{Character, options are 'Account', 'Catalogue', 'Checklist',
'Flora', 'Herbarium collection', 'Key', 'Red list', 'Report',
'Species Database', 'Survey'}

\item{entity_class}{Character, options are 'Island', 'Island/Mainland',
'Mainland', 'Island Group', 'Island Part'}

\item{native_indicated}{Boolean, whether only lists where native status
is available should be retrieved.}

\item{natural_indicated}{Boolean, whether only lists where natural status
is available should be retrieved.}

\item{end_ref}{Boolean, whether only lists where endemism at the reference
level is available should be retrieved.}

\item{end_list}{Boolean, whether only lists where endemism at the list level
is available should be retrieved.}

\item{suit_geo}{Boolean, whether only suitable polygons should be retrieved.}

\item{complete_taxon}{Boolean, default TRUE.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}

\item{list_set}{default NULL. If not, a vector of identification numbers of
GIFT checklists.}

\item{taxonomy}{default NULL. If not, it has to be the taxonomy table (see
GIFT_taxonomy()).}
}
\value{
A data frame with 16 columns.
}
\description{
Retrieve GIFT checklists that fulfill specific criteria.
}
\details{
Here is what each column refers to:

\emph{ref_ID} - Identification number of each reference.\cr
\emph{type} - What type the source is.\cr
\emph{subset} - What information regarding the status of species is
available.\cr
\emph{native_indicated} - Whether native status of species is available in
 the source.\cr
\emph{natural_indicated} - Whether naturalized status of species is
available in the source.\cr
\emph{end_ref} - Whether endemism information is available in the source.\cr
\emph{restricted} - Whether the access to this reference is restricted.\cr
\emph{taxon_ID} - Identification number of species.\cr
\emph{list_ID} - Identification number of each list.\cr
\emph{end_list} - Whether endemism information is available in the list.\cr
\emph{entity_ID} - Identification number of the polygon of the list.\cr
\emph{geo_entity} - Name of the location.\cr
\emph{suit_geo} - Is the polygon suitable.\cr
\emph{entity_class} - Type of polygon.\cr
\emph{entity_type} - Name of the location.\cr
\emph{taxon_name} - Name of the group of taxa available.
}
\examples{
\dontrun{
ex <- GIFT_checklist_conditional(taxon_name = "Embryophyta", 
ref_included = c("all", "native", "native and naturalized",
"native and historically introduced", "endangered",
"endemic", "naturalized", "other subset")[1:4],
type_ref = c("Account", "Catalogue", "Checklist","Flora",
"Herbarium collection", "Key", "Red list", "Report", "Species Database",
 "Survey"),
 entity_class = c("Island", "Island/Mainland", "Mainland", "Island Group",
 "Island Part"),
 native_indicated = FALSE, natural_indicated = FALSE, end_ref = FALSE,
 end_list = FALSE, suit_geo = TRUE, complete_taxon = TRUE,
 list_set = NULL, taxonomy = NULL)

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist_raw()]
}
