% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_checklist_raw.R
\name{GIFT_checklist_raw}
\alias{GIFT_checklist_raw}
\title{GIFT checklists}
\usage{
GIFT_checklist_raw(
  ref_ID = NULL,
  list_ID = NULL,
  namesmatched = FALSE,
  taxon_name = "Tracheophyta",
  floristic_group = "all",
  list_set = NULL,
  taxonomy = NULL,
  GIFT_version = NULL,
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{ref_ID}{A vector defining the IDs of the references to retrieve.
`NULL` by default.}

\item{list_ID}{A vector defining the IDs of the lists to retrieve.
`NULL` by default. These lists are retrieved in addition to the lists 
contained in the references in `ref_ID`.}

\item{namesmatched}{Boolean. FALSE by default, set to TRUE if you want the
original species name as they came in the references as well as details on
the taxonomic harmonization.}

\item{taxon_name}{Character.}

\item{floristic_group}{Character string among these options:
'all', 'native', 'naturalized', 'endemic_list', 'endemic_ref'.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}
}
\value{
A data frame with 13 columns, if namesmatched = FALSE.
}
\description{
Raw checklist, to combine with other functions
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_checklist_raw(list_ID = c(1,5))
ex2 <- GIFT_checklist_raw(list_ID = c(1,5), namesmatched = TRUE)
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist()]
}
