% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_checklist_raw.R
\name{GIFT_checklist_raw}
\alias{GIFT_checklist_raw}
\title{GIFT checklists}
\usage{
GIFT_checklist_raw(
  list_ID = NULL,
  taxonid = 1,
  namesmatched = FALSE,
  floristic_group = NULL,
  api = "http://gift.uni-goettingen.de/api/extended/index.php"
)
}
\arguments{
\item{list_ID}{A vector defining the ID of the lists to retrieve.
`NULL` by default, in that case, every list from GIFT is retrieved.}

\item{taxonid}{Integer.}

\item{namesmatched}{Boolean: do you want the full species name.}

\item{floristic_group}{NULL or character among these options: 'native',
'naturalized', 'endemic_list', 'endemic_ref'.}

\item{api}{character string defining from which API the data will be retrieved.}
}
\value{
A data frame with 12 columns, if namesmatched = FALSE.
}
\description{
Raw checklist, to combine with other functions
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_checklist_raw(list_ID = c(1,5), api = api)
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist_raw()]
}
