% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_env.R
\name{GIFT_env}
\alias{GIFT_env}
\title{Environmental data for GIFT checklists}
\usage{
GIFT_env(
  entity_ID = NULL,
  miscellaneous = "area",
  rasterlayer = NULL,
  sumstat = "mean",
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{entity_ID}{A vector defining the ID of the lists to retrieve.
`NULL` by default, in that case, every list from GIFT is retrieved.}

\item{miscellaneous}{character vector or list defining the miscellaneous
data to retrieve.}

\item{rasterlayer}{character vector or list defining the raster
data to retrieve.}

\item{sumstat}{Vector or list indicating the desired summary statistics out 
of c("min", "q05", "q10", "q20", "q25", "q30", "q40", "med", "q60", "q70", 
"q75", "q80", "q90", "q95", "max", "mean", "sd", "modal", "unique_n", "H", 
"n") used to aggregate the information coming from the raster layers. If 
sumstat is a vector, the same summary statistics are used for all raster 
layers. If sumstat is a list, the first element defines the summary 
statistics for the first raster layer, the second for the second and so
on.\cr
\strong{Important note} Some summary statistics may not be informative
depending on the environmental layer you ask for. For example, it is not
relevant to retrieve the mean of soil classes for a polygon. The mode or
Shannon index are more suitable in that case.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}
}
\value{
A data frame with the environmental values per polygon (entity_ID).
}
\description{
Retrieve environmental data associated to each GIFT checklists.
They can come as rasters or shapefiles (miscellaneous)
}
\details{
The columns of the data.frame are the following:

\emph{entity_ID} - Identification number of the polygon\cr
\emph{geo_entity} - Name of the polygon\cr
The other columns relate to the environmental variables the user asked for.
}
\examples{
\dontrun{
ex <- GIFT_env(miscellaneous = "perimeter")
ex <- GIFT_env(entity_ID = c(1,5), miscellaneous = c("perimeter", "biome"))

ex <- GIFT_env(entity_ID = c(1,5),
               miscellaneous = c("perimeter", "biome"),
               rasterlayer = c("mn30_grd", "wc2.0_bio_30s_01"),
               sumstat = "mean")

ex <- GIFT_env(entity_ID = c(1,5),
               miscellaneous = c("perimeter", "biome"),
               rasterlayer = c("mn30_grd", "wc2.0_bio_30s_01"),
               sumstat = c("mean", "med"))

ex <- GIFT_env(entity_ID = c(1,5),
               miscellaneous = c("perimeter", "biome"),
               rasterlayer = c("mn30_grd", "wc2.0_bio_30s_01"),
               sumstat = list(c("mean", "med"), "max"))

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_env_meta_misc()] and [GIFT::GIFT_env_meta_raster()]
}
