% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_env.R
\name{GIFT_env}
\alias{GIFT_env}
\title{Environmental data for GIFT checklists}
\usage{
GIFT_env(
  entity_ID = NULL,
  miscellaneous = "area",
  rasterlayer = NULL,
  sumstat = "mean",
  api = "http://gift.uni-goettingen.de/api/extended/index.php"
)
}
\arguments{
\item{entity_ID}{A vector defining the ID of the lists to retrieve.
`NULL` by default, in that case, every list from GIFT is retrieved.}

\item{miscellaneous}{character vector or list defining the miscellaneous
data to retrieve.}

\item{rasterlayer}{character vector or list defining the raster
data to retrieve..}

\item{sumstat}{Vector with the summary statistics to aggregate the
information coming from the raster layers.}

\item{api}{character string defining from which API the data will be retrieved.}
}
\value{
data frame.
}
\description{
Retrieve environmental data associated to each GIFT checklists.
They can come as rasters or shapefiles (miscellaneous)
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_env(miscellaneous = "perimeter")
ex <- GIFT_env(entity_ID = c(1,5), miscellaneous = c("perimeter", "biome"))

ex <- GIFT_env(entity_ID = c(1,5),
               miscellaneous = c("perimeter", "biome"),
               rasterlayer = c("mn30_grd", "wc2.0_bio_30s_01"),
               sumstat = "mean")

ex <- GIFT_env(entity_ID = c(1,5),
               miscellaneous = c("perimeter", "biome"),
               rasterlayer = c("mn30_grd", "wc2.0_bio_30s_01"),
               sumstat = c("mean", "med"))

ex <- GIFT_env(entity_ID = c(1,5),
               miscellaneous = c("perimeter", "biome"),
               rasterlayer = c("mn30_grd", "wc2.0_bio_30s_01"),
               sumstat = list(c("mean", "med"), "max"))

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist_raw()]
}
