% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_env_meta_misc.R
\name{GIFT_env_meta_misc}
\alias{GIFT_env_meta_misc}
\title{Metadata for the environmental miscellaneous variables in GIFT}
\usage{
GIFT_env_meta_misc(
  api = "https://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{api}{character string defining from which API the data will be
retrieved.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data frame with 6 columns.
}
\description{
Retrieve the metadata of miscellaneous environmental layers accessible in
GIFT
}
\details{
Here is what each column refers to:

\emph{dataset} - Name of the source dataset.\cr
\emph{variable} - Name of the environmental layer.\cr
\emph{description}- Description.\cr
\emph{unit} - Unit.\cr
\emph{num} - Whether the environmental layer is numeric or not.\cr
\emph{ref_long} - Full reference to cite when using an environmental layer.
}
\examples{
\dontrun{
ex <- GIFT_env_meta_misc()
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=GIFT_env]{GIFT_env()}}
}
