% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_individual.R
\name{GIFT_individual}
\alias{GIFT_individual}
\title{The main function for conditional fine-mapping for in transcriptome-wide association studies with individual-level data}
\usage{
GIFT_individual(
  X,
  Y,
  Zx,
  Zy,
  pindex,
  max_iterin = 1000,
  epsin = 1e-04,
  Cores = 1
)
}
\arguments{
\item{X}{complex gene expression matrix, each column is the standardized specific expression.}

\item{Y}{standarized trait vector.}

\item{Zx}{standardized cis-genotype matrix in eQTL data.}

\item{Zy}{standardized cis-genotype matrix in GWAS data.}

\item{pindex}{a vector with each element represents the number of cis-SNPs for each gene.}

\item{max_iterin}{The maximum iteration, which can be determined by users.}

\item{epsin}{The convergence tolerance of the absolute value of the difference  between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{Cores}{The number of cores used in analysis. If the number of cores is greater than 1, analysis will perform with fast parallel computing. The function mclapply() depends on another R package "parallel" in Linux.}
}
\value{
A list of estimated parameters including the p values for the gene-based test. 
\item{causal_effect}{The estimates of causal effect for each gene in a specific region}
\item{gene_based_test_pvalue}{The p values for each gene by the gene-based test}
}
\description{
GIFT_individual applies a likelihood-based approach, accounting for the correlated cis-SNPs and genes in a region
}
