% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_phylogeny.R
\name{GIFT_phylogeny}
\alias{GIFT_phylogeny}
\title{Phylogeny of the species in GIFT}
\usage{
GIFT_phylogeny(
  taxon_name = "Tracheophyta",
  as_newick = TRUE,
  api = "https://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{taxon_name}{Character string indicating the taxonomic group
of interest corresponding to the node labels in the phylogeny.}

\item{as_newick}{Boolean, whether you want the phylogeny to be in
Newick tree format (TRUE) or in table format (FALSE). TRUE by default.}

\item{api}{character string defining the API from which the data will be
retrieved.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the latest stable
version.}
}
\value{
A data frame with 5 columns or a tree object.
}
\description{
Retrieve a phylogeny of the plant species available in GIFT.
}
\details{
Here is what each column refers to:

\emph{taxon_label} - Name of the taxonomic group\cr
\emph{work_ID} - Standardized species name IDs for the species at the tips
of the tree\cr
\emph{edge_length} - Edge length\cr
\emph{lft} - Left border of a given taxon in the Newick sequence\cr
\emph{rgt} - Right border of a given taxon in the Newick sequence\cr
}
\examples{
\dontrun{
ex <- GIFT_phylogeny(taxon_name = "Tracheophyta", as_newick = TRUE)
ex2 <- GIFT_phylogeny(taxon_name = "Tracheophyta", as_newick = FALSE)
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=GIFT_checklist]{GIFT_checklist()}}
}
