% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_shape.R
\name{GIFT_shape}
\alias{GIFT_shape}
\title{GIFT shape}
\usage{
GIFT_shape(
  entity_ID = NULL,
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{entity_ID}{A vector defining the IDs of the regions.}

\item{api}{character string defining from which API the data will be
retrieved.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A spatial data.frame with 21 columns.
}
\description{
Get shapefile of GIFT regions for selected entities
}
\details{
Here is the detail of each column:

\emph{entity_ID} - Identification number of the polygon\cr
\emph{geo_entity} - Name of the polygon\cr
\emph{point_x} - Longitude of the centroid of the polygon\cr
\emph{point_y} - Latitude of the centroid of the polygon\cr
\emph{area} - Area in km2 of the polygon\cr
\emph{x_min} - Minimum longitude of the polygon\cr
\emph{x_max} - Maximum longitude of the polygon\cr
\emph{y_min} - Minimum latitude of the polygon\cr
\emph{y_max} - Maximum latitude of the polygon\cr
\emph{entity_class} - Class of the polygon\cr
\emph{entity_type} - Type of the entity\cr
\emph{polygon_source} - Source of the polygon\cr
\emph{geometry} - Geometry column from sf
}
\examples{
\dontrun{
ex <- GIFT_shape(entity_ID = c(677, 200))
plot(st_geometry(ex), col = ex$entity_ID)
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
 Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
 https://doi.org/10.1111/jbi.13623
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=GIFT_plot]{GIFT_plot()}}, \code{\link[=GIFT_env]{GIFT_env()}}
}
