% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_spatial.R
\name{GIFT_spatial}
\alias{GIFT_spatial}
\title{Environmental data for GIFT checklists}
\usage{
GIFT_spatial(
  shp = NULL,
  coordinates = NULL,
  overlap = "centroid_inside",
  entity_ID = NULL,
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{shp}{Shapefile provided by the user.}

\item{coordinates}{Custom set of coordinates. The format is a two columns,
the first one being longitudes and the second being latitudes. If 4
coordinates are given, the function assumes that these are the four corners
of a bounding box.}

\item{overlap}{character vector or list defining the raster
data to retrieve..}

\item{entity_ID}{List of entity_ID to retrieve.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be retrieved.}
}
\value{
data frame with list of entity_ID.
}
\description{
Retrieve environmental data associated to each GIFT checklists.
They can come as rasters or shapefiles (miscellaneous)
}
\details{
Blabla.
}
\examples{
\dontrun{
data("med")
ex <- GIFT_spatial(shp = med, overlap = "centroid_inside")
ex2 <- GIFT_spatial(shp = med, overlap = "extent_intersect")
ex3 <- GIFT_spatial(shp = med, overlap = "shape_intersect")
ex4 <- GIFT_spatial(shp = med, overlap = "shape_inside")

custom_point <- cbind(9.9, 51)
ex5 <- GIFT_spatial(coordinates = custom_point,
overlap = "extent_intersect")

custom_extent <- cbind(c(-13, -18), c(27.5, 29.3))
ex6 <- GIFT_spatial(coordinates = custom_extent,
overlap = "extent_intersect")

custom_polygon <- cbind(c(-18, -16.9, -13, -13, -18, -18),
c(29.3, 33, 29.3, 27.5, 27.5, 29.3))
ex7 <- GIFT_spatial(coordinates = custom_polygon,
overlap = "extent_intersect")

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklists()]
}
