% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_spatial.R
\name{GIFT_spatial}
\alias{GIFT_spatial}
\title{Environmental data for GIFT checklists}
\usage{
GIFT_spatial(
  shp = NULL,
  coordinates = NULL,
  overlap = "centroid_inside",
  api = "http://gift.uni-goettingen.de/api/extended/index.php"
)
}
\arguments{
\item{shp}{Shapefile provided by the user.}

\item{coordinates}{Coordinates (point, extent box or polygon) provided by
the user.}

\item{overlap}{character vector or list defining the raster
data to retrieve..}

\item{api}{character string defining from which API the data will be retrieved.}
}
\value{
data frame with list of entity_ID.
}
\description{
Retrieve environmental data associated to each GIFT checklists.
They can come as rasters or shapefiles (miscellaneous)
}
\details{
Blabla.
}
\examples{
\dontrun{
data("med")
ex <- GIFT_spatial(shp = med, overlap = "centroid_inside")

ex_extent <- c(-15, 40, 30, 45)
ex_extent <- c(120, 124, -13, -12)

ex <- GIFT_spatial(shp = med, overlap = "extent_intersect")
ex2 <- GIFT_spatial(shp = med, overlap = "shape_intersect")
ex3 <- GIFT_spatial(shp = med, overlap = "shape_inside")

# Example with custom polygon: first and last row need to be repeated
# Example with matrix(, byrow = TRUE) or own matrix with repeated x and y
# Basically: matrix with two columns: first X, second Y

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklists()]
}
