% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_species_distribution.R
\name{GIFT_species_distribution}
\alias{GIFT_species_distribution}
\title{GIFT species distribution}
\usage{
GIFT_species_distribution(
  genus = "Fagus",
  epithet = "sylvatica",
  namesmatched = FALSE,
  remove_overlap = FALSE,
  area_th_island = 0,
  area_th_mainland = 100,
  overlap_th = 0.1,
  by_ref_ID = FALSE,
  GIFT_version = "latest",
  api = "http://gift.uni-goettingen.de/api/extended/"
)
}
\arguments{
\item{genus}{Character string corresponding to the genus of the species 
of interest.}

\item{epithet}{Character string corresponding to the epithet of the species 
of interest.}

\item{namesmatched}{Boolean. FALSE by default, set to TRUE if you want to 
look for the species not only in the standardized species names but also 
in the original species names as they came in the original resources.}

\item{remove_overlap}{a boolean stating whether you want to
retrieve checklists that overlap or not.}

\item{area_th_island}{A number stating from which surface the smallest
overlapping polygon is kept. By default set to 0 square kilometer
(meaning that by default the smallest islands will be conserved).}

\item{area_th_mainland}{When two polygons overlap, the smallest or the
biggest one can be kept. When the surface of the smallest polygon exceeds
this number, the smallest polygon is kept. Otherwise, we keep the bigger
one. Set by default 100 square-kilometers.}

\item{overlap_th}{A number ranging from 0 to 1, indicating at what
percentage of overlap, partially overlapping polygons should be kept.}

\item{by_ref_ID}{logical indicating whether the removal of overlapping
regions shall be applied by ref_ID only. Note that regions overlapping with
other regions from the same resource will be removed even if there are other
references available for those regions.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}

\item{api}{character string defining from which API the data will be
retrieved.}
}
\value{
List with two data frames: the checklist with species and the list of ID.
}
\description{
Retrieve the distribution of one species from GIFT checklits.
}
\details{
Here is the detail of each data.frame and their columns:
}
\examples{
\dontrun{
ex <- GIFT_species_distribution()
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_species_lookup()]
}
