% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_species_lookup.R
\name{GIFT_species_lookup}
\alias{GIFT_species_lookup}
\title{Species list in GIFT}
\usage{
GIFT_species_lookup(
  genus = "",
  epithet = "",
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{genus}{character string defining the genus name to be looked for.}

\item{epithet}{character string defining the specific epithet to be looked
for.}

\item{api}{character string defining from which API the data will be
retrieved.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data frame with 24 columns.
}
\description{
Retrieve all name matching information for one taxonomic name. All results 
are returned, where the name is either found in the unstandardized or 
taxonomically standardized names.
}
\details{
Here is what each column refers to:
'orig_ID' - 
'orig_genus'- 
'name_ID'-
}
\examples{
\dontrun{
ex <- GIFT_species_lookup(genus = "Fagus", epithet = "sylvatica")
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_checklist()]
}
