% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_summary.R
\name{GIFT_summary}
\alias{GIFT_summary}
\title{The main function for conditional fine-mapping for in transcriptome-wide association studies with summary-level data}
\usage{
GIFT_summary(
  Zscore1,
  Zscore2,
  Sigma1,
  Sigma2,
  n1,
  n2,
  gene,
  pindex,
  R = NULL,
  maxiter = 100,
  tol = 1e-03,
  pleio = 0,
  ncores = 1,
  in_sample_LD = F,
  filter = T
)
}
\arguments{
\item{Zscore_1}{Zscore matrix of the cis-SNP effect size matrix, each column for one specific gene from eQTL data.}

\item{Zscore_2}{Zscore vector of the cis-SNP effect size vector from GWAS data.}

\item{Sigma1}{LD matrix from eQTL data.}

\item{Sigma2}{LD matrix from GWAS data, both Sigma1 and Sigma2 are often the same from the reference panel.}

\item{n1}{Sample size of eQTL data.}

\item{n2}{Sample size of GWAS data.}

\item{gene}{The gene names vector.}

\item{pindex}{A vector with each element represents the number of cis-SNPs for each gene.}

\item{R}{Estimated correlation matrix of gene expressions.}

\item{maxiter}{The maximum iteration, which can be determined by users.}

\item{tol}{The convergence tolerance of the absolute value of the difference  between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{pleio}{The option of controlling the pleiotropy, which can be determined by users. If pleio is set to 0, analysis will perform without controlling any SNP; If pleio is set to 1, analysis will perform  controlling the top SNP; If pleio is set to 2, analysis will perform controlling the top two SNPs.}

\item{ncores}{The number of cores used in analysis. If the number of cores is greater than 1, analysis will perform with fast parallel computing. The function mclapply() depends on another R package "parallel" in Linux.}

\item{in_sample_LD}{A logical value, which represents whether in-sample LD was used. If in-sample LD was not used, the LD matrix is regularized to be (1-s1)*Sigma1+s1*E and (1-s2)*Sigma2+s2*E, where s1 and s2 are estimated by estimate_s_rss in susieR, and E is an identity matrix. A grid search is performed over the range from 0.1 to 1 if the estimation does not work. The function estimate_s_rss() depends on another R package "susieR".}

\item{filter}{A logical value, which represents whether filter the SNP with GWAS P>0.05 when the GWAS sample size over 100,000. This step will improve the computational speed.}
}
\value{
A data frame including the causal effect estimates and p values for the gene-based test.
}
\description{
GIFT_summary applies a likelihood-based approach, accounting for the correlated cis-SNPs and genes in a region
}
