% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_taxgroup.R
\name{GIFT_taxgroup}
\alias{GIFT_taxgroup}
\title{GIFT taxgroup}
\usage{
GIFT_taxgroup(
  work_ID = NULL,
  taxon_lvl = c("family", "order", "higher_lvl")[1],
  return_ID = FALSE,
  GIFT_version = NULL,
  api = "http://gift.uni-goettingen.de/api/extended/",
  taxonomy = NULL,
  species = NULL
)
}
\arguments{
\item{work_ID}{A vector defining the IDs of the species to retrieve taxonomic 
groups for. `Null` by default.}

\item{taxon_lvl}{taxonomic level to retrieve names for. "family" by default. 
Check `GIFT_taxonomy()` for available levels. In addition to the available 
levels one can put "higher_lvl" to retrieve the higher level groups 
"Anthocerotophyta", "Marchantiophyta", "Bryophyta", "Lycopodiophyta", 
"Monilophyta", "Gymnospermae", and "Angiospermae".}

\item{return_ID}{logical indicating whether to give back taxon_IDs instead of 
names.}

\item{api}{character string defining from which API the data will be retrieved.}

\item{taxonomy}{option to supply taxonomy object here if loaded already to 
avoid double loading. For internal use within GIFT functions. If `NULL` 
(default) taxonomy will be loaded within this function.}

\item{species}{option to supply species names object here if loaded already to 
avoid double loading. For internal use within GIFT functions. If `NULL` 
(default) species will be loaded within this function.}

\item{higher_lvl}{logical indicating whether to retrieve the higher level 
groups (Angiospermae, Gymnospermae, Pteridophyta, ...).}
}
\value{
A data frame with 13 columns, if namesmatched = FALSE.
}
\description{
Assign taxonomic groups of various hierarchical level to species from GIFT 
(work_ID)
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_taxgroup(work_ID = c(1:5), taxon_lvl = "family")
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_taxonomy()]
}
