% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits.R
\name{GIFT_traits}
\alias{GIFT_traits}
\title{GIFT_traits}
\usage{
GIFT_traits(
  trait_IDs = "",
  agreement = 0.66,
  bias_ref = TRUE,
  bias_deriv = TRUE,
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{trait_IDs}{a character string indicating which trait you want to
retrieve. Traits must belong to the available list of traits.}

\item{agreement}{Percentage of resources that agree on an aggregated trait
value, entries below this threshold will be omitted.}

\item{bias_ref}{When FALSE, exclude entries that are only based on a
resource that potentially introduces a bias (e.g. a resource only including
trees).}

\item{bias_deriv}{When FALSE, exclude entries that are only based on a
derivation that potentially introduces a bias (e.g. all phanerophytes being
woody but some life forms being ambiguous).}

\item{api}{Character string with the API.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A long-format data frame with 6 columns: `trait_ID`, `work_ID`, `species`,
`trait_value`, `agreement` and `references`.
}
\description{
Retrieve specific trait values.
}
\details{
Here is the detail of each column:

\emph{trait_ID} - Identification number of the trait\cr
\emph{work_ID} - Identification number of the taxonomically harmonized
 species\cr
\emph{species} - Species name\cr
\emph{trait_value} - Value of the trait\cr
\emph{agreement} - Agreement score between the different sources for that
trait value\cr
\emph{references} - ref_ID from which we got the trait information
}
\examples{
\dontrun{
wood <- GIFT_traits(trait_IDs = c("1.1.1", "1.2.1"), agreement = 0.66,
bias_ref = FALSE, bias_deriv = FALSE)
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()]
}
