% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits_meta.R
\name{GIFT_traits_meta}
\alias{GIFT_traits_meta}
\title{GIFT_traits_meta}
\usage{
GIFT_traits_meta(
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{api}{Character string corresponding to the API.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data frame with 10 columns.
}
\description{
Retrieve metadata of the functional traits.
}
\details{
Here is what each column refers to:
'Lvl1' - First level of the trait classification.
'Category'- Name of the first level of classification.
'Lvl2'- Second level of the trait classification.
'Trait1'- ame of the second level of classification.
'Lvl3' - Identification number of the trait.
'Trait2' - Trait name.
'Units' - Trait unit.
'type'- Trait type.
'comment'- Comment.
'count' - How many entries for that traits are in the database.
}
\examples{
\dontrun{
ex <- GIFT_traits_meta()
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits()]
}
