% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits_meta.R
\name{GIFT_traits_meta}
\alias{GIFT_traits_meta}
\title{GIFT_traits_meta}
\usage{
GIFT_traits_meta(
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = NULL
)
}
\arguments{
\item{api}{Character string corresponding to the API.}
}
\value{
data frame with metadata about traits available in GIFT.
}
\description{
Retrieve metadata of the functional traits.
}
\details{
Blabla.
}
\examples{
\dontrun{
ex <- GIFT_traits_meta()
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits()]
}
