% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits_raw.R
\name{GIFT_traits_raw}
\alias{GIFT_traits_raw}
\title{GIFT_traits_raw}
\usage{
GIFT_traits_raw(
  trait_IDs = "",
  derived = TRUE,
  bias_ref = TRUE,
  bias_deriv = TRUE,
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = NULL
)
}
\arguments{
\item{trait_IDs}{a character string indicating which traits you want to
retrieve. Traits must belong to the available list of traits. See 
GIFT_traits_meta()}

\item{derived}{include logically derived traits.}

\item{bias_ref}{Traits desired.}

\item{bias_deriv}{Traits desired.}

\item{api}{Character string with the API.}
}
\value{
A
}
\description{
Retrieve non-aggregated trait values at the level of the bibliographic 
references and un-standardized species names in GIFT.
}
\details{
Blabla.
}
\examples{
\dontrun{

}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()] and [GIFT::GIFT_traits()]
}
