% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits_raw.R
\name{GIFT_traits_raw}
\alias{GIFT_traits_raw}
\title{GIFT_traits_raw}
\usage{
GIFT_traits_raw(
  trait_IDs = "",
  derived = TRUE,
  bias_ref = TRUE,
  bias_deriv = TRUE,
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{trait_IDs}{a character string indicating which traits you want to
retrieve. Traits must belong to the available list of traits. See 
GIFT_traits_meta()}

\item{derived}{include logically derived traits.}

\item{bias_ref}{When FALSE, exclude entries that are only based on a
resource that potentially introduces a bias (e.g. a resource only including
trees).}

\item{bias_deriv}{When FALSE, exclude entries that are only based on a
derivation that potentially introduces a bias (e.g. all phanerophytes being
woody but some life forms being ambiguous).}

\item{api}{Character string with the API.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data.frame with 26 columns.
}
\description{
Retrieve non-aggregated trait values at the level of the bibliographic 
references and un-standardized species names in GIFT.
}
\details{
Here is the detail of each column:
trait_derived_ID -
ref_ID - Identification number of the reference
orig_ID - Identification number of the species, as it came in the source
trait_ID - Identification number of the trait
trait_value - Value of the trait
derived - Is the trait value derived from another information (e.g.
phanerophytes are woody)
bias_deriv - Is the derivation potentially introducing a bias
name_ID - Identification number of the species before being resolved
cf_genus -
genus - Genus of the species
cf_species -
aff_species -
species_epithet - Epithet of the species
subtaxon - Sub-taxon name
author - Author who described the species
matched - Was the species name matched in the taxonomic backbone
epithetscore - Matching score for the epithet
overallscore - Overall matching score
resolved - Was the species name resolved in the taxonomic backbone
service - Taxonomic backbone used for taxonomic harmonization
work_ID - Identification number of the taxonomically harmonized species
bias_ref - Is the resource potentially introducing a bias
species - Species name
geo_entity_ref - Name of the region of the reference
ref_long - Full reference to cite
}
\examples{
\dontrun{
succulence_carnivory <- GIFT_traits_raw(trait_IDs = c("4.10.1", "4.16.1"))
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()] and [GIFT::GIFT_traits()]
}
