% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFT_traits_raw.R
\name{GIFT_traits_raw}
\alias{GIFT_traits_raw}
\title{GIFT_traits_raw}
\usage{
GIFT_traits_raw(
  trait_IDs = "",
  derived = TRUE,
  bias_ref = TRUE,
  bias_deriv = TRUE,
  api = "http://gift.uni-goettingen.de/api/extended/",
  GIFT_version = "latest"
)
}
\arguments{
\item{trait_IDs}{a character string indicating which traits you want to
retrieve. Traits must belong to the available list of traits. See 
GIFT_traits_meta()}

\item{derived}{include logically derived traits.}

\item{bias_ref}{When FALSE, exclude entries that are only based on a
resource that potentially introduces a bias (e.g. a resource only including
trees).}

\item{bias_deriv}{When FALSE, exclude entries that are only based on a
derivation that potentially introduces a bias (e.g. all phanerophytes being
woody but some life forms being ambiguous).}

\item{api}{Character string with the API.}

\item{GIFT_version}{character string defining the version of the GIFT
database to use. The function retrieves by default the most up-to-date
version.}
}
\value{
A data.frame with 26 columns.
}
\description{
Retrieve non-aggregated trait values at the level of the bibliographic 
references and un-standardized species names in GIFT.
}
\details{
Here is the detail of each column:

\emph{trait_derived_ID} - \cr
\emph{ref_ID} - Identification number of the reference\cr
\emph{orig_ID} - Identification number of the species, as it came in the
 source\cr
\emph{trait_ID} - Identification number of the trait\cr
\emph{trait_value} - Value of the trait\cr
\emph{derived} - Is the trait value derived from another information (e.g.
phanerophytes are woody)\cr
\emph{bias_deriv} - Is the derivation potentially introducing a bias\cr
\emph{name_ID} - Identification number of the species before being
resolved\cr
\emph{cf_genus} - \cr
\emph{genus} - Genus of the species\cr
\emph{cf_species} - \cr
\emph{aff_species} - \cr
\emph{species_epithet} - Epithet of the species\cr
\emph{subtaxon} - Sub-taxon name\cr
\emph{author} - Author who described the species\cr
\emph{matched} - Was the species name matched in the taxonomic backbone\cr
\emph{epithetscore} - Matching score for the epithet\cr
\emph{overallscore} - Overall matching score\cr
\emph{resolved} - Was the species name resolved in the taxonomic
backbone\cr
\emph{service} - Taxonomic backbone used for taxonomic harmonization\cr
\emph{work_ID} - Identification number of the taxonomically harmonized
species\cr
\emph{bias_ref} - Is the resource potentially introducing a bias\cr
\emph{species} - Species name\cr
\emph{geo_entity} _ref - Name of the region of the reference\cr
\emph{ref_long} - Full reference to cite
}
\examples{
\dontrun{
succulence_carnivory <- GIFT_traits_raw(trait_IDs = c("4.10.1", "4.16.1"))
}

}
\references{
Weigelt, P, König, C, Kreft, H. GIFT – A Global Inventory of Floras and
     Traits for macroecology and biogeography. J Biogeogr. 2020; 47: 16– 43.
     https://doi.org/10.1111/jbi.13623
}
\seealso{
[GIFT::GIFT_traits_meta()] and [GIFT::GIFT_traits()]
}
